+++ backend/app/Database/Migrations/2026-03-09-000012_CreateGoodsReceiptsTable.php
@@ -12,7 +12,7 @@ class CreateGoodsReceiptsTable extends Migration
     {
         $this->forge->addField([
             'id' => [
-                'type'           => 'INT',
+                'type'           => 'BIGINT',
                 'constraint'     => 11,
                 'unsigned'       => true,
                 'auto_increment' => true,
@@ -24,19 +24,19 @@ class CreateGoodsReceiptsTable extends Migration
                 'comment'    => '進貨單號，格式 GR-YYYYMMDD-NNNN',
             ],
             'purchase_order_id' => [
-                'type'       => 'INT',
+                'type'           => 'BIGINT',
                 'constraint' => 11,
                 'unsigned'   => true,
                 'null'       => false,
             ],
             'warehouse_id' => [
-                'type'       => 'INT',
+                'type'           => 'BIGINT',
                 'constraint' => 11,
                 'unsigned'   => true,
                 'null'       => false,
             ],
             'received_by' => [
-                'type'       => 'INT',
+                'type'           => 'BIGINT',
                 'constraint' => 11,
                 'unsigned'   => true,
                 'null'       => false,
