+++ backend/app/Database/Migrations/2026-03-09-000029_CreateWarehousesTable.php
@@ -12,7 +12,7 @@ class CreateWarehousesTable extends Migration
     {
         $this->forge->addField([
             'id' => [
-                'type'           => 'INT',
+                'type'           => 'BIGINT',
                 'constraint'     => 11,
                 'unsigned'       => true,
                 'auto_increment' => true,
